chore(contentful/dynamo): rename many files to kebab name
eslint plugin to enforce kebab filenames
