+++ packages/botonic-plugin-contentful/package.json
@@ -15,8 +15,8 @@
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
   "dependencies": {
-    "@botonic/core": "^0.9.0-beta.1",
-    "@botonic/react": "^0.9.0-beta.1",
+    "@botonic/core": "0.9.0-beta.1",
+    "@botonic/react": "0.9.0-beta.2",
     "contentful": "^7.6.0",
     "escape-string-regexp": "^2.0.0",
     "moment": "^2.24.0",
@@ -36,6 +36,7 @@
     "eslint": "^5.16.0",
     "eslint-config-oclif": "^3.1.0",
     "eslint-config-prettier": "^4.1.0",
+    "eslint-plugin-filenames": "^1.3.2",
     "eslint-plugin-no-null": "^1.0.2",
     "eslint-plugin-prettier": "^3.0.1",
     "eslint-plugin-react": "^7.12.4",
