+++ packages/botonic-plugin-dynamodb/package.json
@@ -11,7 +11,7 @@
     "postversion": "git push && git push --tags"
   },
   "name": "@botonic/plugin-dynamodb",
-  "version": "0.9.0",
+  "version": "0.9.0-beta.1",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
   "dependencies": {
@@ -27,6 +27,7 @@
     "@typescript-eslint/parser": "^1.6.0",
     "eslint": "^5.16.0",
     "eslint-config-prettier": "^4.1.0",
+    "eslint-plugin-filenames": "^1.3.2",
     "eslint-plugin-no-null": "^1.0.2",
     "eslint-plugin-prettier": "^3.0.1",
     "jest": "^24.7.1",
